requirements.txt
numpy
opencv-python
pillow
matplotlib
streamlit
pandas
scikit-learn
seaborn

